feature: Apply adapter pattern
