+++ sep490-enterprise/src/main/java/enterprise/services/impl/PaymentServiceImpl.java
@@ -1,12 +1,12 @@
 package enterprise.services.impl;
 
 import commons.springfw.impl.utils.SecurityUtils;
+import enterprise.adapters.payos.payos.PayOSAdapter;
 import enterprise.dtos.PaymentCriteriaDTO;
 import enterprise.entities.CreditPackageEntity;
 import enterprise.entities.EnterpriseEntity;
 import enterprise.entities.PaymentEntity;
 import enterprise.entities.WalletEntity;
-import enterprise.mappers.PaymentMapper;
 import enterprise.repositories.CreditPackageRepository;
 import enterprise.repositories.EnterpriseRepository;
 import enterprise.repositories.PaymentRepository;
@@ -15,21 +15,12 @@ import enterprise.services.PaymentService;
 import green_buildings.commons.api.dto.SearchCriteriaDTO;
 import green_buildings.commons.api.enums.PaymentStatus;
 import green_buildings.commons.api.exceptions.TechnicalException;
-import green_buildings.commons.api.utils.EnumUtil;
-import jakarta.annotation.PostConstruct;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.web.util.UriComponentsBuilder;
-import vn.payos.PayOS;
-import vn.payos.type.CheckoutResponseData;
-import vn.payos.type.ItemData;
-import vn.payos.type.PaymentData;
-import vn.payos.type.PaymentLinkData;
 
 import java.util.UUID;
 
@@ -42,34 +33,10 @@ public class PaymentServiceImpl implements PaymentService {
     private final CreditPackageRepository packageRepo;
     private final EnterpriseRepository enterpriseRepo;
     private final WalletRepository walletRepository;
-    private final PaymentMapper mapper;
-    private final PayOS payOS;
     
-    @Value("${payment.payos.returnPath}")
-    private String returnPath;
-    
-    @Value("${payment.payos.cancelPath}")
-    private String cancelPath;
-    
-    @Value("${spring.application.homepage}")
-    private String homepage;
-    
-    private String returnUrl;
-    private String cancelUrl;
+    private final PayOSAdapter payOSAdapter;
     
     public static final String CREDIT_ITEM = "Credit";
-    
-    @PostConstruct
-    public void setUrls() {
-        returnUrl = UriComponentsBuilder.fromUriString(homepage)
-                                        .path(returnPath)
-                                        .build()
-                                        .toUriString();
-        cancelUrl = UriComponentsBuilder.fromUriString(homepage)
-                                        .path(cancelPath)
-                                        .build()
-                                        .toUriString();
-    }
     
     @Override
     public Page<PaymentEntity> search(SearchCriteriaDTO<PaymentCriteriaDTO> searchCriteria, Pageable pageable) {
@@ -86,16 +53,7 @@ public class PaymentServiceImpl implements PaymentService {
             EnterpriseEntity enterpriseEntity = enterpriseRepo.findById(enterpriseUUID).orElseThrow();
             
             // Build info
-            long orderCode = System.currentTimeMillis();
-            
-            ItemData itemData = buildItemData(creditPackageEntity);
-            PaymentData paymentData = getPaymentData(orderCode, creditPackageEntity, enterpriseEntity, itemData);
-            
-            // Perform PayOS API
-            CheckoutResponseData payOSResult = payOS.createPaymentLink(paymentData);
-            
-            // Store to DB
-            PaymentEntity paymentEntity = preparePaymentEntity(enterpriseEntity, payOSResult, creditPackageEntity);
+            PaymentEntity paymentEntity = payOSAdapter.newPayment(creditPackageEntity, enterpriseEntity);
             
             return payRepo.save(paymentEntity);
         } catch (Exception e) {
@@ -104,34 +62,18 @@ public class PaymentServiceImpl implements PaymentService {
         }
     }
     
-    private PaymentEntity preparePaymentEntity(EnterpriseEntity enterpriseEntity, CheckoutResponseData payOSResult,
-                                                  CreditPackageEntity creditPackageEntity) {
-        PaymentEntity paymentEntity = new PaymentEntity();
-        paymentEntity.setEnterprise(enterpriseEntity);
-        paymentEntity.setStatus(PaymentStatus.PENDING);
-        paymentEntity.setAmount(payOSResult.getAmount());
-        paymentEntity.setNumberOfCredits(creditPackageEntity.getNumberOfCredits());
-        mapper.updatePaymentFromCheckoutData(payOSResult, paymentEntity);
-        return paymentEntity;
-    }
-    
     @Override
     public void updatePaymentInfo(Long orderCode) {
         UUID uuid = SecurityUtils.getCurrentUserEnterpriseId().orElseThrow();
         PaymentEntity paymentEntity = payRepo.findByOrderCodeAndEnterpriseId(orderCode, uuid).orElseThrow();
-        try {
-            PaymentLinkData paymentLinkInformation = payOS.getPaymentLinkInformation(orderCode);
-            PaymentStatus newStatus =
-                    EnumUtil.getCodeFromString(PaymentStatus.class, paymentLinkInformation.getStatus())
-                            .orElseThrow(() -> new TechnicalException("Error in getting payment status from PayOS"));
-            updateWallet(newStatus, paymentEntity, uuid);
-            paymentEntity.setStatus(newStatus);
-            payRepo.save(paymentEntity);
-        } catch (TechnicalException e) {
-            throw e;
-        } catch (Exception e) {
-            throw new TechnicalException(e);
-        }
+        
+        PaymentStatus newStatus = payOSAdapter.getPaymentStatus(orderCode)
+                                              .orElseThrow(() -> new TechnicalException("Error in getting payment status from PayOS"));
+        
+        updateWallet(newStatus, paymentEntity, uuid);
+        paymentEntity.setStatus(newStatus);
+        payRepo.save(paymentEntity);
+        
     }
     
     private void updateWallet(PaymentStatus newStatus, PaymentEntity paymentEntity, UUID uuid) {
@@ -140,31 +82,6 @@ public class PaymentServiceImpl implements PaymentService {
             wallet.deposit(paymentEntity.getNumberOfCredits());
             walletRepository.save(wallet);
         }
-    }
-    
-    private PaymentData getPaymentData(long orderCode,
-                                       CreditPackageEntity creditPackageEntity,
-                                       EnterpriseEntity enterpriseEntity,
-                                       ItemData itemData) {
-        return PaymentData.builder()
-                          .orderCode(orderCode)
-                          .amount((int) creditPackageEntity.getPrice())
-                          .description("Credit purchase") // max 25 chars
-                          .buyerName(enterpriseEntity.getName())
-                          .buyerEmail(enterpriseEntity.getEmail())
-                          .buyerPhone(enterpriseEntity.getHotline())
-                          .item(itemData)
-                          .returnUrl(returnUrl)
-                          .cancelUrl(cancelUrl)
-                          .build();
-    }
-    
-    private ItemData buildItemData(CreditPackageEntity creditPackageEntity) {
-        return ItemData.builder()
-                       .name(CREDIT_ITEM)
-                       .quantity(creditPackageEntity.getNumberOfCredits())
-                       .price(0)
-                       .build();
     }
     
 }
